Refactor plugins into directories and centralize keymaps
