+++ lua/plugins/editing/init.lua
@@ -1,0 +1,6 @@
+return {
+  require("plugins.editing.nvim_gomove"),
+  require("plugins.editing.nvim_treesitter"),
+  require("plugins.editing.toggleterm"),
+  require("plugins.editing.rainbow_delimiters"),
+}
