+++ lua/plugins/git/gitsigns.lua
@@ -1,0 +1,8 @@
+return {
+  "lewis6991/gitsigns.nvim",
+  event = { "BufReadPre", "BufNewFile" },
+  opts = {
+    signs = { add = { text = "│" }, change = { text = "│" }, delete = { text = "⎺" } },
+    current_line_blame = true,
+  },
+}
