+++ lua/plugins/git/init.lua
@@ -1,0 +1,3 @@
+return {
+  require("plugins.git.gitsigns"),
+}
