+++ lua/plugins/lsp/init.lua
@@ -1,0 +1,7 @@
+return {
+  require("plugins.lsp.nvim_cmp"),
+  require("plugins.lsp.mason"),
+  require("plugins.lsp.mason_lspconfig"),
+  require("plugins.lsp.uv"),
+  require("plugins.lsp.nvim_lspconfig"),
+}
