+++ lua/plugins/lsp/mason.lua
@@ -1,0 +1,1 @@
+return { "mason-org/mason.nvim", opts = {} }
