+++ lua/plugins/lsp/mason_lspconfig.lua
@@ -1,0 +1,7 @@
+return {
+  "mason-org/mason-lspconfig.nvim",
+  version = "1.*",
+  opts = {
+    ensure_installed = { "pyright", "lua_ls" },
+  },
+}
