+++ lua/plugins/navigation/init.lua
@@ -1,0 +1,4 @@
+return {
+  require("plugins.navigation.smart_splits"),
+  require("plugins.navigation.telescope"),
+}
