+++ lua/plugins/navigation/telescope.lua
@@ -1,0 +1,7 @@
+return {
+  "nvim-telescope/telescope.nvim",
+  dependencies = { "nvim-lua/plenary.nvim" },
+  keys = require("core.keymaps.telescope").keys,
+  cmd = "Telescope",
+  opts = true,
+}
