+++ lua/plugins/ui/init.lua
@@ -1,0 +1,8 @@
+return {
+  require("plugins.ui.cyberdream"),
+  require("plugins.ui.transparent"),
+  require("plugins.ui.lualine"),
+  require("plugins.ui.nvim_tree"),
+  require("plugins.ui.mini"),
+  require("plugins.ui.which_key"),
+}
