+++ lua/plugins/ui/mini.lua
@@ -1,0 +1,6 @@
+return {
+  "echasnovski/mini.nvim",
+  config = function()
+    require("mini.icons").setup()
+  end,
+}
